made search bar into one component
