+++ frontend/src/components/SearchPage/index.js
@@ -1,0 +1,7 @@
+const SearchPage = () => {
+    return (
+        <div>hello from searhc page</div>
+    )
+}
+
+export default SearchPage;
